interpretability

+++ interpretability.py
@@ -53,7 +53,7 @@ def convert_to_img(args, rgb_img, activation_map):
             activation_frame = Image.fromarray(activation_map[f, t]).resize((width, height), PIL.Image.NEAREST)
 
             activation_frame_dst = np.array(Image.new("HSV", activation_frame.size))
-            hue = (float(f) / num_features) * 360
+            hue = int((float(f) / num_features) * 360)
             print(f"f: {f}, h: {hue}")
             activation_frame_dst[..., 0] = hue
             activation_frame_dst = np.array(Image.fromarray(activation_frame_dst).convert("RGBA"))
